requirements.txt
alembic==1.7.7
aniso8601==9.0.1
attrs==21.4.0
click==8.1.2
Flask==2.1.1
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.1
jsonschema==4.4.0
Mako==1.2.0
MarkupSafe==2.1.1
pyrsistent==0.18.1
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.35
Werkzeug==2.1.1
